Version 3_Lines 15 - 21_9-28-2015
I understand, after reading the final version of the poem and the additional texts Professor Kirschenbaum posted, that only the 7 upside down lines are part of the poem. The text that appears at the top of the page is not part of the transcription. 
Here is the finalized version of the correct last lines of the Shelly's poem "My list William, thouh in whom".